refactored collections code into module
